Fix 16F1459 erase bug in uploader8.py
On the 16F1459, program memory is addressed in words rather than bytes.
The uploader code works in bytes, so MEMEND and MEMSTART are defined as
the byte equivalents of board.memstart and board.memend. However, these
new variables are not used when calculating the number of blocks to
erase. The available memory in words is instead divided by the erase
block size in bytes. This error means that blocks beyond address 0x1280
are never erased, causing problems when a program is written to these
blocks.

This change correctly uses MEMEND and MEMSTART to calculate the number
of blocks to erase. This fixes the bug on 16F1459, but I can't be sure
this does not have unintended effects on other 8-bit MCUs as I don't
have any other Pinguino's to test on.