Implement processor-specific USB RAM section
The USB5 section is hardcoded as 0500-05FFh, but there's a ToDo comment
about correcting this for PIC16F145x and PIC18F1xk50. This commit
implements this todo by checking board.proc and setting the appropriate
memory range.

However, it may be cleaner and more future-proof to store the range in
the board object (e.g. board.usbmemstart and board.usbmemend) rather
than hardcoding the ranges in pinguino_tools.py as this commit does.